# Vectorize ProbabilisticMap.LastIndexOfAny